🧪 AY0H1F task: isolate drift recovery e2e
